fix: Allow to pass double values as touch event arguments (#372)

+++ app/src/main/java/io/appium/uiautomator2/model/api/touch/appium/TouchEventParams.java
@@ -3,9 +3,9 @@ package io.appium.uiautomator2.model.api.touch.appium;
 import io.appium.uiautomator2.model.api.ElementModel;
 
 public class TouchEventParams extends ElementModel {
-    public Integer x;
-    public Integer y;
-    public Integer duration;
+    public Double x;
+    public Double y;
+    public Double duration;
 
     public TouchEventParams() {}
 }
